MyRatClassesPElmsReps: rewritten to avoid bug (?) in partition backtrack
This changes MyRatClassesPElmsReps to use DuplicateFreeList on a list
of rational classes, instead of using IsConjugate. The resulting code
is much easier to understand, and, importantly, it manages to compute
the classes of AutomorphismGroup(SmallGroup(2187,9302)) in seconds,
while the old code took forever (I aborted after a few hours), hanging
somewhere in the partition backtrack code.

The resulting code is unfortunately a bit slower (maybe 5%) in cases
were the old code took reasonable amounts of time). But all in all,
I think this is a good tradeoff.

In the future, we might investigate using Ferret's partition backtrack
instead.